Moved post status logic out of MainActivity
