+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/MainPresenter.java
@@ -4,6 +4,12 @@ import android.os.Message;
 
 import androidx.annotation.NonNull;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -12,11 +18,13 @@ import edu.byu.cs.tweeter.client.backgroundTask.GetFollowingCountTask;
 import edu.byu.cs.tweeter.client.backgroundTask.IsFollowerTask;
 import edu.byu.cs.tweeter.client.cache.Cache;
 import edu.byu.cs.tweeter.client.model.service.FollowService;
+import edu.byu.cs.tweeter.client.model.service.StatusService;
 import edu.byu.cs.tweeter.client.model.service.UserService;
 import edu.byu.cs.tweeter.model.domain.AuthToken;
+import edu.byu.cs.tweeter.model.domain.Status;
 import edu.byu.cs.tweeter.model.domain.User;
 
-public class MainPresenter implements UserService.Observer, FollowService.Observer {
+public class MainPresenter implements UserService.Observer, FollowService.Observer, StatusService.Observer {
 
     public interface View {
         void logoutUser();
@@ -68,6 +76,12 @@ public class MainPresenter implements UserService.Observer, FollowService.Observ
         new FollowService().followUnfollow(authToken, selectedUser, wasFollowing, this);
     }
 
+    public void postStatus(String post) throws ParseException {
+        Status newStatus = new Status(post, Cache.getInstance().getCurrUser(),
+                getFormattedDateTime(), parseURLs(post), parseMentions(post));
+        new StatusService().postStatus(authToken, newStatus, this);
+    }
+
     public void updateSelectedUserFollowingAndFollowers() {
         // TODO: Utilize this?
         ExecutorService executor = Executors.newFixedThreadPool(2);
@@ -79,10 +93,80 @@ public class MainPresenter implements UserService.Observer, FollowService.Observ
         getFollowingCount();
     }
 
+    public String getFormattedDateTime() throws ParseException {
+        SimpleDateFormat userFormat = new SimpleDateFormat("yyyy-MM-dd hh:mm:ss");
+        SimpleDateFormat statusFormat = new SimpleDateFormat("MMM d yyyy h:mm aaa");
+
+        return statusFormat.format(userFormat.parse(LocalDate.now().toString() + " " + LocalTime.now().toString().substring(0, 8)));
+    }
+
+    public List<String> parseURLs(String post) {
+        List<String> containedUrls = new ArrayList<>();
+        for (String word : post.split("\\s")) {
+            if (word.startsWith("http://") || word.startsWith("https://")) {
+
+                int index = findUrlEndIndex(word);
+
+                word = word.substring(0, index);
+
+                containedUrls.add(word);
+            }
+        }
+
+        return containedUrls;
+    }
+
+    public List<String> parseMentions(String post) {
+        List<String> containedMentions = new ArrayList<>();
+
+        for (String word : post.split("\\s")) {
+            if (word.startsWith("@")) {
+                word = word.replaceAll("[^a-zA-Z0-9]", "");
+                word = "@".concat(word);
+
+                containedMentions.add(word);
+            }
+        }
+
+        return containedMentions;
+    }
+
+    public int findUrlEndIndex(String word) {
+        if (word.contains(".com")) {
+            int index = word.indexOf(".com");
+            index += 4;
+            return index;
+        } else if (word.contains(".org")) {
+            int index = word.indexOf(".org");
+            index += 4;
+            return index;
+        } else if (word.contains(".edu")) {
+            int index = word.indexOf(".edu");
+            index += 4;
+            return index;
+        } else if (word.contains(".net")) {
+            int index = word.indexOf(".net");
+            index += 4;
+            return index;
+        } else if (word.contains(".mil")) {
+            int index = word.indexOf(".mil");
+            index += 4;
+            return index;
+        } else {
+            return word.length();
+        }
+    }
+
     @Override
     public void handleSuccess(User user, AuthToken authToken) {
         view.clearInfoMessage();
         view.logoutUser();
+    }
+
+    @Override
+    public void handleSuccess(List<Status> statuses, boolean hasMorePages) {
+        view.clearInfoMessage();
+        view.displayInfoMessage("Successfully Posted!");
     }
 
     @Override
